First pass at handing odd edges from switch to host.
If we see the PR type in the switch connection output these are (it seems) connections between switches and hosts being reported when we would only expect switch to switch connections. For now just ignore these. We need to come back and validate this to handle it better, and get some metadata about the link (e.g., bandwidth, name).